# server/application-server/openapi.yaml
openapi: 3.1.0
info:
  title: Helios API
  description: API documentation for the Helios application server.
  contact:
    name: Turker Koc
    email: turker.koc@tum.de
  license:
    name: MIT License
    url: https://github.com/ls1intum/Helios/blob/main/LICENSE
  version: 0.0.1
servers:
- url: /
  description: Default Server URL
paths:
...
  /api/tags:
    get:
      tags:
      - tag-controller
      operationId: getAllTags
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TagInfoDto"
    put:
      tags:
      - tag-controller
      operationId: createTag
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagCreateDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagInfoDto"
  /api/tags/{name}/markworking:
    put:
      tags:
      - tag-controller
      operationId: markWorking
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
  /api/tags/{name}/markbroken:
    put:
      tags:
      - tag-controller
      operationId: markBroken
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
...
  /api/tags/{name}:
    get:
      tags:
      - tag-controller
      operationId: getTagByName
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagDetailsDto"
  /api/tags/newcommits/{name}:
    get:
      tags:
      - tag-controller
      operationId: getCommitsSinceLastTag
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommitsSinceTagDto"
...
  /api/branches/repository/{repoId}/branch:
    get:
      tags:
      - branch-controller
      operationId: getBranchByRepositoryIdAndName
      parameters:
      - name: repoId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: name
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BranchDetailsDto"
...
components:
  schemas:
    TagCreateDto:
      type: object
      properties:
        name:
          type: string
        commitSha:
          type: string
        branchName:
          type: string
        createdByLogin:
          type: string
      required:
      - commitSha
      - createdByLogin
      - name
    TagInfoDto:
      type: object
      properties:
        name:
          type: string
        commitSha:
          type: string
        branchName:
          type: string
...
    EnvironmentDeployment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        url:
          type: string
        state:
          type: string
          enum:
          - PENDING
          - WAITING
          - SUCCESS
          - ERROR
          - FAILURE
          - IN_PROGRESS
          - QUEUED
          - INACTIVE
          - UNKNOWN
        statusesUrl:
          type: string
        sha:
          type: string
        ref:
          type: string
        task:
          type: string
        tagName:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
      - id
      - ref
      - sha
      - statusesUrl
      - task
      - url
    EnvironmentDto:
      type: object
      properties:
        repository:
          $ref: "#/components/schemas/RepositoryInfoDto"
        id:
          type: integer
          format: int64
        name:
          type: string
        locked:
          type: boolean
        url:
          type: string
        htmlUrl:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        enabled:
          type: boolean
        installedApps:
          type: array
          items:
            type: string
        description:
          type: string
        serverUrl:
          type: string
        latestDeployment:
          $ref: "#/components/schemas/EnvironmentDeployment"
        lockedBy:
          type: string
        lockedAt:
          type: string
          format: date-time
      required:
      - id
      - name
    RepositoryInfoDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        nameWithOwner:
          type: string
        description:
          type: string
        htmlUrl:
          type: string
      required:
      - htmlUrl
      - id
      - name
      - nameWithOwner
...
    BranchInfoDto:
      type: object
      properties:
        name:
          type: string
        commitSha:
          type: string
        aheadBy:
          type: integer
          format: int32
        behindBy:
          type: integer
          format: int32
        isDefault:
          type: boolean
        isProtected:
          type: boolean
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          $ref: "#/components/schemas/UserInfoDto"
        repository:
          $ref: "#/components/schemas/RepositoryInfoDto"
      required:
      - commitSha
      - name
    CommitInfoDto:
      type: object
      properties:
        sha:
          type: string
        author:
          $ref: "#/components/schemas/UserInfoDto"
        message:
          type: string
        authoredAt:
          type: string
          format: date-time
        repository:
          $ref: "#/components/schemas/RepositoryInfoDto"
      required:
      - sha
    DeploymentDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        repository:
          $ref: "#/components/schemas/RepositoryInfoDto"
        url:
          type: string
        state:
          type: string
          enum:
          - PENDING
          - WAITING
          - SUCCESS
          - ERROR
          - FAILURE
          - IN_PROGRESS
          - QUEUED
          - INACTIVE
          - UNKNOWN
        statusesUrl:
          type: string
        sha:
          type: string
        ref:
          type: string
        task:
          type: string
        environment:
          $ref: "#/components/schemas/EnvironmentDto"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
      - environment
      - id
      - ref
      - sha
      - statusesUrl
      - task
      - url
    TagDetailsDto:
      type: object
      properties:
        name:
          type: string
        commit:
          $ref: "#/components/schemas/CommitInfoDto"
        branch:
          $ref: "#/components/schemas/BranchInfoDto"
        deployments:
          type: array
          items:
            $ref: "#/components/schemas/DeploymentDto"
        markedWorkingBy:
          type: array
          items:
            $ref: "#/components/schemas/UserInfoDto"
        markedBrokenBy:
          type: array
          items:
            $ref: "#/components/schemas/UserInfoDto"
        createdBy:
          $ref: "#/components/schemas/UserInfoDto"
        createdAt:
          type: string
          format: date-time
      required:
      - commit
      - createdAt
      - createdBy
      - deployments
      - markedBrokenBy
      - markedWorkingBy
      - name
    UserInfoDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        login:
          type: string
        avatarUrl:
          type: string
        name:
          type: string
        htmlUrl:
          type: string
      required:
      - avatarUrl
      - htmlUrl
      - id
      - login
      - name
    CommitsSinceTagDto:
      type: object
      properties:
        commitsLength:
          type: integer
          format: int32
        commits:
          type: array
          items:
            $ref: "#/components/schemas/CommitInfoDto"
      required:
      - commits
      - commitsLength
...
    BranchDetailsDto:
      type: object
      properties:
        name:
          type: string
        commitSha:
          type: string
        aheadBy:
          type: integer
          format: int32
        behindBy:
          type: integer
          format: int32
        isDefault:
          type: boolean
        isProtected:
          type: boolean
        tagName:
          type: string
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          $ref: "#/components/schemas/UserInfoDto"
        repository:
          $ref: "#/components/schemas/RepositoryInfoDto"
      required:
      - commitSha
      - name